Update permute to use size_t instead of int.

+++ libs/Einsums/TensorAlgebra/src/Permute.cpp
@@ -20,88 +20,88 @@
 
 namespace einsums::tensor_algebra::detail {
 
-void permute(int const *perm, int const dim, float const alpha, float const *A, int const *sizeA, float const beta, float *B) {
+void permute(int const *perm, int const dim, float const alpha, float const *A, size_t const *sizeA, float const beta, float *B) {
     auto plan =
         hptt::create_plan(perm, dim, alpha, A, sizeA, nullptr, beta, B, nullptr, hptt::ESTIMATE, omp_get_max_threads(), nullptr, true);
     plan->execute();
 }
 
-void permute(int const *perm, int const dim, double const alpha, double const *A, int const *sizeA, double const beta, double *B) {
+void permute(int const *perm, int const dim, double const alpha, double const *A, size_t const *sizeA, double const beta, double *B) {
     auto plan =
         hptt::create_plan(perm, dim, alpha, A, sizeA, nullptr, beta, B, nullptr, hptt::ESTIMATE, omp_get_max_threads(), nullptr, true);
     plan->execute();
 }
 
-void permute(int const *perm, int const dim, std::complex<float> const alpha, std::complex<float> const *A, int const *sizeA,
+void permute(int const *perm, int const dim, std::complex<float> const alpha, std::complex<float> const *A, size_t const *sizeA,
              std::complex<float> const beta, std::complex<float> *B) {
     auto plan =
         hptt::create_plan(perm, dim, alpha, A, sizeA, nullptr, beta, B, nullptr, hptt::ESTIMATE, omp_get_max_threads(), nullptr, true);
     plan->execute();
 }
 
-void permute(int const *perm, int const dim, std::complex<double> const alpha, std::complex<double> const *A, int const *sizeA,
+void permute(int const *perm, int const dim, std::complex<double> const alpha, std::complex<double> const *A, size_t const *sizeA,
              std::complex<double> const beta, std::complex<double> *B) {
     auto plan =
         hptt::create_plan(perm, dim, alpha, A, sizeA, nullptr, beta, B, nullptr, hptt::ESTIMATE, omp_get_max_threads(), nullptr, true);
     plan->execute();
 }
 
-void permute(int const *perm, int const dim, float const alpha, float const *A, int const *sizeA, int const *offsetA, int const *outerSizeA,
-             float const beta, float *B, int const *offsetB, int const *outerSizeB) {
+void permute(int const *perm, int const dim, float const alpha, float const *A, size_t const *sizeA, size_t const *offsetA, size_t const *outerSizeA,
+             float const beta, float *B, size_t const *offsetB, size_t const *outerSizeB) {
     auto plan = hptt::create_plan(perm, dim, alpha, A, sizeA, outerSizeA, offsetA, beta, B, outerSizeB, offsetB, hptt::ESTIMATE,
                                   omp_get_max_threads(), nullptr, true);
     plan->execute();
 }
 
-void permute(int const *perm, int const dim, double const alpha, double const *A, int const *sizeA, int const *offsetA,
-             int const *outerSizeA, double const beta, double *B, int const *offsetB, int const *outerSizeB) {
+void permute(int const *perm, int const dim, double const alpha, double const *A, size_t const *sizeA, size_t const *offsetA,
+             size_t const *outerSizeA, double const beta, double *B, size_t const *offsetB, size_t const *outerSizeB) {
     auto plan = hptt::create_plan(perm, dim, alpha, A, sizeA, outerSizeA, offsetA, beta, B, outerSizeB, offsetB, hptt::ESTIMATE,
                                   omp_get_max_threads(), nullptr, true);
     plan->execute();
 }
 
-void permute(int const *perm, int const dim, std::complex<float> const alpha, std::complex<float> const *A, int const *sizeA,
-             int const *offsetA, int const *outerSizeA, std::complex<float> const beta, std::complex<float> *B, int const *offsetB,
-             int const *outerSizeB) {
+void permute(int const *perm, int const dim, std::complex<float> const alpha, std::complex<float> const *A, size_t const *sizeA,
+             size_t const *offsetA, size_t const *outerSizeA, std::complex<float> const beta, std::complex<float> *B, size_t const *offsetB,
+             size_t const *outerSizeB) {
     auto plan = hptt::create_plan(perm, dim, alpha, A, sizeA, outerSizeA, offsetA, beta, B, outerSizeB, offsetB, hptt::ESTIMATE,
                                   omp_get_max_threads(), nullptr, true);
     plan->execute();
 }
 
-void permute(int const *perm, int const dim, std::complex<double> const alpha, std::complex<double> const *A, int const *sizeA,
-             int const *offsetA, int const *outerSizeA, std::complex<double> const beta, std::complex<double> *B, int const *offsetB,
-             int const *outerSizeB) {
+void permute(int const *perm, int const dim, std::complex<double> const alpha, std::complex<double> const *A, size_t const *sizeA,
+             size_t const *offsetA, size_t const *outerSizeA, std::complex<double> const beta, std::complex<double> *B, size_t const *offsetB,
+             size_t const *outerSizeB) {
     auto plan = hptt::create_plan(perm, dim, alpha, A, sizeA, outerSizeA, offsetA, beta, B, outerSizeB, offsetB, hptt::ESTIMATE,
                                   omp_get_max_threads(), nullptr, true);
     plan->execute();
 }
 
-void permute(int const *perm, int const dim, float const alpha, float const *A, int const *sizeA, int const *offsetA, int const *outerSizeA,
-             int const innerStrideA, float const beta, float *B, int const *offsetB, int const *outerSizeB, int const innerStrideB) {
+void permute(int const *perm, int const dim, float const alpha, float const *A, size_t const *sizeA, size_t const *offsetA, size_t const *outerSizeA,
+             size_t const innerStrideA, float const beta, float *B, size_t const *offsetB, size_t const *outerSizeB, size_t const innerStrideB) {
     auto plan = hptt::create_plan(perm, dim, alpha, A, sizeA, outerSizeA, offsetA, innerStrideA, beta, B, outerSizeB, offsetB, innerStrideB,
                                   hptt::ESTIMATE, omp_get_max_threads(), nullptr, true);
     plan->execute();
 }
 
-void permute(int const *perm, int const dim, double const alpha, double const *A, int const *sizeA, int const *offsetA,
-             int const *outerSizeA, int const innerStrideA, double const beta, double *B, int const *offsetB, int const *outerSizeB,
-             int const innerStrideB) {
+void permute(int const *perm, int const dim, double const alpha, double const *A, size_t const *sizeA, size_t const *offsetA,
+             size_t const *outerSizeA, size_t const innerStrideA, double const beta, double *B, size_t const *offsetB, size_t const *outerSizeB,
+             size_t const innerStrideB) {
     auto plan = hptt::create_plan(perm, dim, alpha, A, sizeA, outerSizeA, offsetA, innerStrideA, beta, B, outerSizeB, offsetB, innerStrideB,
                                   hptt::ESTIMATE, omp_get_max_threads(), nullptr, true);
     plan->execute();
 }
 
-void permute(int const *perm, int const dim, std::complex<float> const alpha, std::complex<float> const *A, int const *sizeA,
-             int const *offsetA, int const *outerSizeA, int const innerStrideA, std::complex<float> const beta, std::complex<float> *B,
-             int const *offsetB, int const *outerSizeB, int const innerStrideB) {
+void permute(int const *perm, int const dim, std::complex<float> const alpha, std::complex<float> const *A, size_t const *sizeA,
+             size_t const *offsetA, size_t const *outerSizeA, size_t const innerStrideA, std::complex<float> const beta, std::complex<float> *B,
+             size_t const *offsetB, size_t const *outerSizeB, size_t const innerStrideB) {
     auto plan = hptt::create_plan(perm, dim, alpha, A, sizeA, outerSizeA, offsetA, innerStrideA, beta, B, outerSizeB, offsetB, innerStrideB,
                                   hptt::ESTIMATE, omp_get_max_threads(), nullptr, true);
     plan->execute();
 }
 
-void permute(int const *perm, int const dim, std::complex<double> const alpha, std::complex<double> const *A, int const *sizeA,
-             int const *offsetA, int const *outerSizeA, int const innerStrideA, std::complex<double> const beta, std::complex<double> *B,
-             int const *offsetB, int const *outerSizeB, int const innerStrideB) {
+void permute(int const *perm, int const dim, std::complex<double> const alpha, std::complex<double> const *A, size_t const *sizeA,
+             size_t const *offsetA, size_t const *outerSizeA, size_t const innerStrideA, std::complex<double> const beta, std::complex<double> *B,
+             size_t const *offsetB, size_t const *outerSizeB, size_t const innerStrideB) {
     auto plan = hptt::create_plan(perm, dim, alpha, A, sizeA, outerSizeA, offsetA, innerStrideA, beta, B, outerSizeB, offsetB, innerStrideB,
                                   hptt::ESTIMATE, omp_get_max_threads(), nullptr, true);
     plan->execute();
